chore(clients/go): moved worker into pkg
