+++ clients/go/Makefile
@@ -7,6 +7,6 @@ install-deps:
 test:
 	go test -v ./pkg/commands/
 	go test -v ./pkg/entities/
-	go test -v ./worker/
+	go test -v ./pkg/worker/
 	go test -v ./tests/
 	go test -v ./zbc/
